pyproject.toml
[project]
name = "gpu-comparison-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-mgmt-compute>=34.0.0",
    "boto3>=1.36.16",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "requests>=2.32.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "aiohttp>=3.11.12",
]
